CURLOPT_RESOLVE: support "*" for port number
Allow resolves for a hostname independent of port number, to a fixed
address. Resolves to a fixed port number have priority.

Also works if the hostname itself is using "*" to catch all hostnames.

A key to this change is to no longer use the port number from the DNS
struct stored in the cache when connecting since it is not the correct
one when a wildcard port is used for CURLOPT_RESOLVE.

Removes TODO item 1.8

Verified by test 1488 and 1489