configure: fix `SystemConfiguration` detection
Before this patch, detection failed due to this error when compiling
the detection snippet:
```
/Applications/Xcode_15.3.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX.sdk/usr/include/TargetConditionals.h:140:50: error: missing binary operator before token "("
  140 | #if !defined(__has_extension) || !__has_extension(define_target_os_macros)
      |                                                  ^
```
Ref: https://github.com/curl/curl/actions/runs/9821817534/job/27117929218#step:6:1079

It occured with gcc-11 when combined with macOS SDK 14.4 and 14.5
(default SDKs in Xcode 15.3 and 15.4 respectively). It did not happen
with earlier releases.

It was interesting is that despite the failure, `lib/macos.c` compiled
fine while using the same SDK header.

Turns out that including `sys/types.h` makes it compile fine also in
the test snippet.

Cherry-picked from #14097
Closes #xxxxx