fix doh cleanup
When removing an easy handle that had DoH sub-easy handles going,
those were not removed from the multi handle. Their memory was
reclaimed on curl_easy_cleanup() of the owning handle, but multi
still had them in their list.

Add `Curl_doh_close()` and `Curl_doh_cleanup()` as common point
for handling the DoH resource management. Use the `multi` present
in the doh handles (if so), for removal, as the `data->multi`
might already have been NULLed at this time.

- refs #14207