cmake: drop `if(PKG_CONFIG_FOUND)` around `pkg_check_modules()`
The oldest cmake supported by curl is 3.7.0, which already has a guard
inside `pkg_check_modules()`. The advantage of leaving that to CMake is
that it will define/reset all variables, while the manual guard doesn't
do this and also leave `NETTLE_FOUND` undefined.

Delete the single use of this check in the recently added NETTLE patch
by accident. And re-introduce this later everywhere if we find it useful after
more evaluation.

Follow-up to 669ce42275635dc1f881dab3dfc9a55c9ab49b21 #14285
Closes #14309