http/2, better rate limiting of downloads
Set the initial stream window size to 64KB and increase that
to the 10MB we used to start with on the first server reply,
unless a rate limit is in effect.

Continously monitory changes to the transfers rate limit and
adjust the stream window size accordingly. `max_recv_speed`
is a transfer propert that can be changed during processing
by a callback.