aws_sigv4: Fix canon order for headers with same prefix
If a request containing two headers that have equivalent prefixes
(ex. "x-amz-meta-test:test" and "x-amz-meta-test-two:test2") AWS
expects the header with the shorter name to come first. The previous
implementation used `strcmp` on the full header. Using the example,
this would result in a comparison between the ':' and '-' chars and
sort "x-amz-meta-test-two" before "x-amz-meta-test", which produces
a different "StringToSign" than the one calculated by AWS.