multi and share debug handling
Debug functions can be installed on a multi and share that
provide tracing of events related to that multi or share.
Tracing is only in effect when the multi/share is set to VERBOSE.

This allows multi/share traces unrelated to a specific transfer
or on libcurl internal handles. Currently this is used in connection
shutdown handling.

Unused, these new options have no impact on existing verbosity and
tracing of transfers. When activated, the multi/share debug function
will be used when a transfer is set to verbose, but has no debug
function of its own.

- new multi options: CURLMOPT_VEROSE, CURLMOPT_DEBUGDATA,
  CURLMOPT_DEBUGFUNCTION
- new share options: CURLSHOPT_VERBOSE, CURLSHOPT_DEBUGDATA,
  CURLSHOPT_DEBUGFUNCTION
- added use of these features in the curl command line. On verbose
  runs, the operations of connection shutdown now become visible in
  the trace.

Introduce trace feat 'multi' for messages related to the operations
of transfers and connections.

- add CURL_TRC_M() as macro for emitting such
- use CURL_TRC_M with prefix '[STATE]' for transfer state
  changes instead of DEBUGF
- use CURL_TRC_M with prefix '[CONN-CACHE]' for connection cache
  related messages
- use CURL_TRC_M in some other places instead of DEBUGF
- add documentation