mbedtls: no longer use MBEDTLS_SSL_VERIFY_OPTIONAL
With mbedTLS if the minimum version of TLS is set to 1.3,
MBEDTLS_SSL_VERIFY_OPTIONAL is not available in client mode. See:
https://github.com/Mbed-TLS/mbedtls/blob/2ca6c285/library/ssl_tls.c#L1357
Also, there might be plans to remove it completely in future mbedTLS
versions.

So, switch to always using MBEDTLS_SSL_VERIFY_REQUIRED. If verifypeer
or verifyhost are disabled the corresponding error flags are cleared
in the verify callback function. That is also where verification
errors are logged.