go back to not calling find_package_handle_standard_args() for pkg-config detections
Making the call doesn't give anything useful. The only difference
is that it will defined the `<original_case>_FOUND` variable over
the fully uppercase version. And also omits the standard 'Found' log
messages.

On the other hand it will show a misleading error message when the
dependency is not found, but saying that the `_INCLUDE_DIRS` and
`_LIBRARIES` vars are missing, while in fact we use `_INCLUDE_DIR`
and `_LIBRARY` for that.

Re-add the log message manually for pkg-config detections to smoothen
out that difference.