restore STATUS for Features/Protocols
Without status the output goes to stderr, and goes red in CMake GUI.
This was not intended, the goal was to align these lines exactly with
the curl -V output to make it easy to compare visually. This doesn't
seem doable with CMake.

Partial revert of acbc6b703f6b0ee568d053f6f2565fbc107b5fd3 #14197