formdata: prefer _fseeki64 when available
It syncs Windows builds with each other.

Before this patch autotools mingw-w64 2.0.0 and later builds detected and
used fseeko(). Cmake did not with the same toolchain.

Another solution is to enable the fseeko check for mingw-w64 in cmake,
but its and extra check and it's also non-trivial to figure out what
fseeko gets translated to at the end and if it has proper 64-bit always.
It's PPd to this in local tests with mingw-w64 12.0.0:
```
fseeko64(stream, (off_t)offset, whence);
```