Pipewait handling, do not wait if there are only single-use connections
If the connection reuse sees a connection to the desired destination
that is not connected yet, it will wait for the connect if the transfer
allows it. Once connected, the connection may allow mutiplexing and
the transfer will then just piggy-back to it.

However, if there are established connection to destination that
do not offer multiplexing and no other connections that do, it seems
safe to assume that the server will not suddenly negotiate multiplexing
on other, pending connections. Do not wait for a connect to come through
in this case and start a new connection right away.