Skip EOS read when download done
When we downloaded all we wanted,
and we did not want a response body,
and no Trailer: has been announced,
and the receive gives EAGAIN,
do not hang around unnecessarily.

Some servers are buggy in HEAD processing and
fail to send the HTTP/2 EOS. Since we do not
need any more data, end the request right there.
This will cause us to send a RST_STREAM to the
server.