connection info update
Update IP related information at the connection and the
transfer in two places only: once the filter chain connects
and when a transfer is added to a connection. The latter
only updates on reuse when the filters already are connected.

The only user of that information before a full connect is
the HAProxy filter. Add cfilter CF_QUERY_IP_INFO query to let
it find the information from the filters "below".

This solves two issues with the previous version:
- updates where often done twice with the same info
- happy eyeballing filter "forks" could overwrite each others
  updates before the full winner was determined.