add terrible hack to avoid autotools meltdown
At the moment the depends list is assemble space-separated within the
configure script and those spaces replaced with commas before copying
the value to `libcurl.pc`.

But, version-dependent depends value can actually contain spaces:
`openssl >= 3.0.0`.

So, I did the next logical thing and tried assembling the list with
commas. Turns out this makes autotools getting confused, picking up
commas within strings as its own markers (or something) and generating
corrupt shell code.

To avoid it, I used a replacement character to represent space and
substitute that too before copying the value to `libcurl.pc`.

I was wondering if the spaces are required around the `>=` operator.
There is no spec and all examples come with them. Testing revealed the
spaces are actually required. (as of pkg-config 0.29.2)