drop the warning message
If it turns out to be important, it's can be added, but I'd only do it
for RTSP, _if_ RTSP wasn't already disabled. Otherwise it's just
a permanent warning with this option enabled.

trurl CI builds use this option via curl-for-win, to reduce size. Such
warning does not seem useful in that context.