shutdown send, proceed normally on timeout
When ending an FTP upload, we shut down the connection gracefully,
since the server should be notified we had send all bytes. Mostly,
this is a NOP without TLS involved. With TLS, close-notify messages
should be exchanged.

As reported in #14843, not all servers seem to do that. Since it
is the server's responsiblity to check it has received everything,
we just log the timeout and proceed as if everything is fine.

In the receive direction, we still fail the transfer if the server
does not shut down its direction properly.