Optionally, let a transfer ignore errors during shutdown
A transfer can flag on setup that a connection (send or receive)
needs to be shut down at the end of the request. If this shutdown
fails, the request fails.

Add an additional flag that ignores a shutdown error, to not fail
the request when this happens.

This is used for FTP uploads. FTP servers have no MUST obligations
to close DATA connections at the end of an upload immediately. They
might delay that for their own reasons. It's not good practise, but
some servers do. The shutdown from curl's side will the time out
or may report a missing close-notify from the server. For FTP
uploads, this is not really relevant, as the server will tell the
client on the CONTROL connection if it considered the STORE a
success or not.